+++ addons/io_hubs_addon/__init__.py
@@ -1,12 +1,14 @@
-from .io import gltf_exporter
 from . import (nodes, components)
+from .io import gltf_exporter, gltf_importer, panels
+import bpy
 from . import preferences
+
 bl_info = {
     "name": "Hubs Blender Addon",
     "author": "Mozilla Hubs",
-    "description": "Tools for developing GLTF assets for Mozilla Hubs",
+    "description": "Tools for developing glTF assets for Mozilla Hubs",
     "blender": (3, 1, 2),
-    "version": (1, 0, 0, "dev_build"),
+    "version": (1, 0, 0),
     "location": "",
     "wiki_url": "https://github.com/MozillaReality/hubs-blender-exporter",
     "tracker_url": "https://github.com/MozillaReality/hubs-blender-exporter/issues",
@@ -18,7 +20,9 @@ bl_info = {
 
 def register():
     preferences.register()
+    panels.register()
     gltf_exporter.register()
+    gltf_importer.register()
     nodes.register()
     components.register()
 
@@ -26,17 +30,20 @@ def register():
 def unregister():
     components.unregister()
     nodes.unregister()
+    gltf_importer.unregister()
     gltf_exporter.unregister()
+    panels.unregister()
     preferences.unregister()
 
 
 # called by gltf-blender-io after it has loaded
 
-
 glTF2ExportUserExtension = gltf_exporter.glTF2ExportUserExtension
 glTF2_pre_export_callback = gltf_exporter.glTF2_pre_export_callback
 glTF2_post_export_callback = gltf_exporter.glTF2_post_export_callback
+if bpy.app.version > (3, 0, 0):
+    glTF2ImportUserExtension = gltf_importer.glTF2ImportUserExtension
 
 
 def register_panel():
-    return gltf_exporter.register_export_panel()
+    return panels.register_panels()
