+++ addons/io_hubs_addon/components/definitions/loop_animation.py
@@ -1,12 +1,14 @@
+from ...io.utils import add_hubs_import_component, assign_property
 import bpy
 from bpy.app.handlers import persistent
-from bpy.props import StringProperty, CollectionProperty, IntProperty, BoolProperty, EnumProperty, FloatProperty
+from bpy.props import StringProperty, CollectionProperty, IntProperty, EnumProperty, FloatProperty
 from bpy.types import PropertyGroup, Menu, Operator
 from ..hubs_component import HubsComponent
 from ..types import Category, PanelType, NodeType
 from ..utils import redraw_component_ui
 
 msgbus_owner = None
+
 
 class TrackPropertyType(PropertyGroup):
     name: StringProperty(
@@ -17,11 +19,11 @@ class TrackPropertyType(PropertyGroup):
         name="Track Name",
         description="Track Name",
     )
-    strip_name: StringProperty( # Will only contain data if the track name is generic
+    strip_name: StringProperty(  # Will only contain data if the track name is generic
         name="Strip Name",
         description="Strip Name",
     )
-    action_name: StringProperty( # Will only contain data if the track name is generic
+    action_name: StringProperty(  # Will only contain data if the track name is generic
         name="Action Name",
         description="Action Name",
     )
@@ -31,9 +33,10 @@ class TrackPropertyType(PropertyGroup):
         items=[
             ("object", "Object", "Object"),
             ("shape_key", "Shape Key", "Shape Key")
-            ],
+        ],
         default="object"
     )
+
 
 class Errors():
     _errors = {}
@@ -42,7 +45,8 @@ class Errors():
     def log(cls, track, error_type, error_message, severity='Error'):
         has_error = cls._errors.get(track.track_type+track.name, '')
         if not has_error:
-            cls._errors[track.track_type+track.name] = {'type': error_type, 'message': error_message, 'severity': severity}
+            cls._errors[track.track_type+track.name] = {
+                'type': error_type, 'message': error_message, 'severity': severity}
 
     @classmethod
     def get(cls, track):
@@ -67,13 +71,15 @@ class Errors():
             error_row.scale_y = 0.7
 
             if i == 0:
-                error_row.label(text=f"{error['severity']}: {line}", icon='ERROR')
+                error_row.label(
+                    text=f"{error['severity']}: {line}", icon='ERROR')
             else:
                 error_row.label(text=line, icon='BLANK1')
 
         padding = layout.row(align=False)
         padding.scale_y = 0.2
         padding.label()
+
 
 def register_msgbus():
     global msgbus_owner
@@ -104,6 +110,7 @@ def register_msgbus():
         notify=redraw_component_ui,
     )
 
+
 def unregister_msgbus():
     global msgbus_owner
     if not msgbus_owner:
@@ -112,21 +119,26 @@ def unregister_msgbus():
     bpy.msgbus.clear_by_owner(msgbus_owner)
     msgbus_owner = None
 
+
 @persistent
 def load_post(dummy):
     unregister_msgbus()
     register_msgbus()
+
 
 @persistent
 def undo_redo_post(dummy):
     unregister_msgbus()
     register_msgbus()
 
+
 def is_default_name(track_name):
     return bool(track_name.startswith("NlaTrack") or track_name.startswith("[Action Stash]"))
 
+
 def get_display_name(track_name, strip_name):
     return track_name if not is_default_name(track_name) else f"{track_name} ({strip_name})"
+
 
 def get_strip_name(nla_track):
     try:
@@ -134,14 +146,17 @@ def get_strip_name(nla_track):
     except IndexError:
         return ''
 
+
 def get_action_name(nla_track):
     try:
         return nla_track.strips[0].action.name
     except (IndexError, AttributeError):
         return ''
 
+
 def get_menu_id(nla_track, track_type, display_name):
     return display_name if not is_default_name(nla_track.name) else track_type+display_name
+
 
 def is_unique_action(animation_data, target_nla_track):
     try:
@@ -163,6 +178,7 @@ def is_unique_action(animation_data, target_nla_track):
 
     return True
 
+
 def has_track(tracks_list, nla_track, invalid_track=None):
     strip_name = get_strip_name(nla_track)
     action_name = get_action_name(nla_track)
@@ -180,6 +196,7 @@ def has_track(tracks_list, nla_track, invalid_track=None):
 
     return exists
 
+
 def is_matching_track(nla_track_type, nla_track, track):
     if nla_track_type != track.track_type:
         return False
@@ -190,13 +207,15 @@ def is_matching_track(nla_track_type, nla_track, track):
                 if get_action_name(nla_track) == track.action_name:
                     return True
 
-                Errors.log(track, 'INVALID_ACTION', "The action has changed for this strip/track.\nChoose the track again to update.")
+                Errors.log(track, 'INVALID_ACTION',
+                           "The action has changed for this strip/track.\nChoose the track again to update.")
 
     else:
         if nla_track.name == track.track_name:
             return True
 
     return False
+
 
 def is_useable_nla_track(animation_data, nla_track, track):
     track_name = nla_track.name
@@ -209,16 +228,19 @@ def is_useable_nla_track(animation_data, nla_track, track):
     forbidden_chars = [",", " "]
     if not is_default_name(track_name):
         if any([c for c in forbidden_chars if c in track_name]):
-            Errors.log(track, 'FORBIDDEN_NAME', "Custom track names can't contain commas or spaces.")
+            Errors.log(track, 'FORBIDDEN_NAME',
+                       "Custom track names can't contain commas or spaces.")
             return False
 
     else:
         if any([c for c in forbidden_chars if c in action_name]):
-            Errors.log(track, 'FORBIDDEN_NAME', "Action names can't contain commas or spaces.")
+            Errors.log(track, 'FORBIDDEN_NAME',
+                       "Action names can't contain commas or spaces.")
             return False
 
     if len(nla_track.strips) > 1:
-        Errors.log(track, 'MULTIPLE_STRIPS', "Only one strip is allowed in the track.")
+        Errors.log(track, 'MULTIPLE_STRIPS',
+                   "Only one strip is allowed in the track.")
         return False
 
     if not nla_track.strips:
@@ -226,22 +248,27 @@ def is_useable_nla_track(animation_data, nla_track, track):
         return False
 
     if nla_track.strips[0].mute:
-        Errors.log(track, 'MUTED_STRIP', "The strip is muted and won't export.")
+        Errors.log(track, 'MUTED_STRIP',
+                   "The strip is muted and won't export.")
         return False
 
     if not action_name:
-        Errors.log(track, 'NO_ACTION', "The strip/track doesn't have an action.")
+        Errors.log(track, 'NO_ACTION',
+                   "The strip/track doesn't have an action.")
         return False
 
     if not nla_track.strips[0].action.fcurves:
-        Errors.log(track, 'NO_FCURVES', "The strip/track's action doesn't have any animation and\nwon't be exported.")
+        Errors.log(track, 'NO_FCURVES',
+                   "The strip/track's action doesn't have any animation and\nwon't be exported.")
         return False
 
     if not is_unique_action(animation_data, nla_track):
-        Errors.log(track, 'NON_UNIQUE_ACTION', "This strip/track contains an action that is present in multiple\nstrips/tracks on this object and may not export correctly.", severity="Warning")
+        Errors.log(track, 'NON_UNIQUE_ACTION',
+                   "This strip/track contains an action that is present in multiple\nstrips/tracks on this object and may not export correctly.", severity="Warning")
         return False
 
     return True
+
 
 def is_valid_regular_track(ob, track):
     if ob.animation_data:
@@ -255,6 +282,7 @@ def is_valid_regular_track(ob, track):
     Errors.log(track, 'NOT_FOUND', "Track not found.  Did you mean:")
 
     return False
+
 
 def is_valid_shape_key_track(ob, track):
     if hasattr(ob.data, 'shape_keys') and ob.data.shape_keys and ob.data.shape_keys.animation_data:
@@ -287,15 +315,17 @@ class TracksList(bpy.types.UIList):
                            emboss=False, icon='SHAPEKEY_DATA')
                 split.enabled = False
             else:
-                spacer = '  ' # needed so the menu arrow doesn't intersect with the name
+                spacer = '  '  # needed so the menu arrow doesn't intersect with the name
                 row = split.row(align=False)
                 row.emboss = 'NONE'
                 row.alignment = 'LEFT'
                 row.context_pointer_set('hubs_component', data)
                 row.context_pointer_set('track', item)
-                row.menu(UpdateTrackContextMenu.bl_idname, text=item.name+spacer, icon='ERROR')
+                row.menu(UpdateTrackContextMenu.bl_idname,
+                         text=item.name+spacer, icon='ERROR')
             row = split.row(align=True)
-            row.emboss = 'UI_EMBOSS_NONE_OR_STATUS' if bpy.app.version < (3, 0, 0) else 'NONE_OR_STATUS'
+            row.emboss = 'UI_EMBOSS_NONE_OR_STATUS' if bpy.app.version < (
+                3, 0, 0) else 'NONE_OR_STATUS'
         elif self.layout_type == 'GRID':
             layout.alignment = 'CENTER'
             layout.label(text="", icon_value=icon)
@@ -320,7 +350,6 @@ class UpdateTrack(Operator):
 
     track_type: StringProperty(
         name="Track Type", description="Track Type", default="")
-
 
     def execute(self, context):
         track = context.track
@@ -445,11 +474,13 @@ class UpdateTrackContextMenu(Menu):
                     row.context_pointer_set('track', track)
 
                     update_track = row.operator(UpdateTrack.bl_idname,
-                                         icon='OBJECT_DATA', text=display_name)
+                                                icon='OBJECT_DATA', text=display_name)
                     update_track.name = display_name
                     update_track.track_name = nla_track.name
-                    update_track.strip_name = strip_name if is_default_name(nla_track.name) else ''
-                    update_track.action_name = action_name if is_default_name(nla_track.name) else ''
+                    update_track.strip_name = strip_name if is_default_name(
+                        nla_track.name) else ''
+                    update_track.action_name = action_name if is_default_name(
+                        nla_track.name) else ''
                     update_track.track_type = track_type
 
                     no_tracks = False
@@ -468,11 +499,13 @@ class UpdateTrackContextMenu(Menu):
                     row.context_pointer_set('track', track)
 
                     update_track = row.operator(UpdateTrack.bl_idname,
-                                         icon='SHAPEKEY_DATA', text=display_name)
+                                                icon='SHAPEKEY_DATA', text=display_name)
                     update_track.name = display_name
                     update_track.track_name = nla_track.name
-                    update_track.strip_name = strip_name if is_default_name(nla_track.name) else ''
-                    update_track.action_name = action_name if is_default_name(nla_track.name) else ''
+                    update_track.strip_name = strip_name if is_default_name(
+                        nla_track.name) else ''
+                    update_track.action_name = action_name if is_default_name(
+                        nla_track.name) else ''
                     update_track.track_type = track_type
 
                     no_tracks = False
@@ -505,11 +538,13 @@ class TracksContextMenu(Menu):
 
                 if menu_id not in menu_tracks and not has_track(component_tracks_list, nla_track):
                     add_track = layout.operator(AddTrackOperator.bl_idname,
-                                         icon='OBJECT_DATA', text=display_name)
+                                                icon='OBJECT_DATA', text=display_name)
                     add_track.name = display_name
                     add_track.track_name = nla_track.name
-                    add_track.strip_name = strip_name if is_default_name(nla_track.name) else ''
-                    add_track.action_name = action_name if is_default_name(nla_track.name) else ''
+                    add_track.strip_name = strip_name if is_default_name(
+                        nla_track.name) else ''
+                    add_track.action_name = action_name if is_default_name(
+                        nla_track.name) else ''
                     add_track.track_type = track_type
                     add_track.panel_type = panel_type.value
 
@@ -526,11 +561,13 @@ class TracksContextMenu(Menu):
 
                 if menu_id not in menu_tracks and not has_track(component_tracks_list, nla_track):
                     add_track = layout.operator(AddTrackOperator.bl_idname,
-                                         icon='SHAPEKEY_DATA', text=display_name)
+                                                icon='SHAPEKEY_DATA', text=display_name)
                     add_track.name = display_name
                     add_track.track_name = nla_track.name
-                    add_track.strip_name = strip_name if is_default_name(nla_track.name) else ''
-                    add_track.action_name = action_name if is_default_name(nla_track.name) else ''
+                    add_track.strip_name = strip_name if is_default_name(
+                        nla_track.name) else ''
+                    add_track.action_name = action_name if is_default_name(
+                        nla_track.name) else ''
                     add_track.track_type = track_type
                     add_track.panel_type = panel_type.value
 
@@ -608,7 +645,8 @@ class LoopAnimation(HubsComponent):
     def gather(self, export_settings, object):
         final_track_names = []
         for track in object.hubs_component_loop_animation.tracks_list.values():
-            final_track_names.append(track.track_name if not is_default_name(track.track_name) else track.action_name)
+            final_track_names.append(track.track_name if not is_default_name(
+                track.track_name) else track.action_name)
 
         fps = bpy.context.scene.render.fps/bpy.context.scene.render.fps_base
 
@@ -618,7 +656,6 @@ class LoopAnimation(HubsComponent):
             'startOffset': self.startOffset/fps,
             'timeScale': self.timeScale
         }
-
 
     @staticmethod
     def register():
@@ -637,7 +674,6 @@ class LoopAnimation(HubsComponent):
             bpy.app.handlers.load_post.append(undo_redo_post)
 
         register_msgbus()
-
 
     @staticmethod
     def unregister():
@@ -680,10 +716,13 @@ class LoopAnimation(HubsComponent):
                             track = host.hubs_component_loop_animation.tracks_list.add()
                             strip_name = get_strip_name(nla_track)
                             action_name = get_action_name(nla_track)
-                            track.name = get_display_name(nla_track.name, strip_name)
+                            track.name = get_display_name(
+                                nla_track.name, strip_name)
                             track.track_name = nla_track.name
-                            track.strip_name = strip_name if is_default_name(nla_track.name) else ''
-                            track.action_name = action_name if is_default_name(nla_track.name) else ''
+                            track.strip_name = strip_name if is_default_name(
+                                nla_track.name) else ''
+                            track.action_name = action_name if is_default_name(
+                                nla_track.name) else ''
                             track.track_type = track_type
 
             for ob in bpy.data.objects:
@@ -692,6 +731,23 @@ class LoopAnimation(HubsComponent):
                 if ob.type == 'ARMATURE':
                     for bone in ob.data.bones:
                         migrate_data(ob, bone)
+
+    @classmethod
+    def gather_import(cls, import_settings, blender_object, component_name, component_value):
+        blender_component = add_hubs_import_component(
+            component_name, blender_object)
+
+        for property_name, property_value in component_value.items():
+            if property_name == 'clip':
+                tracks = property_value.split(",")
+                for track_name in tracks:
+                    if not has_track(blender_component.tracks_list, track_name):
+                        track = blender_component.tracks_list.add()
+                        track.name = track_name.strip()
+
+            else:
+                assign_property(import_settings.vnodes, blender_component,
+                                property_name, property_value)
 
 
 def register_module():
