+++ addons/io_hubs_addon/io/gltf_exporter.py
@@ -1,12 +1,6 @@
 import bpy
-from bpy.props import PointerProperty, IntVectorProperty
 from ..components.components_registry import get_components_registry
-from .utils import gather_lightmap_texture_info
-
-hubs_config = {
-    "gltfExtensionName": "MOZ_hubs_components",
-    "gltfExtensionVersion": 4,
-}
+from .utils import gather_lightmap_texture_info, HUBS_CONFIG
 
 if bpy.app.version < (3, 0, 0):
     from io_scene_gltf2.io.exp.gltf2_io_user_extensions import export_user_extensions
@@ -77,11 +71,11 @@ class glTF2ExportUserExtension:
         if not self.properties.enabled or not self.was_used:
             return
 
-        extension_name = hubs_config["gltfExtensionName"]
+        extension_name = HUBS_CONFIG["gltfExtensionName"]
         gltf2_object.extensions[extension_name] = self.Extension(
             name=extension_name,
             extension={
-                "version": hubs_config["gltfExtensionVersion"],
+                "version": HUBS_CONFIG["gltfExtensionVersion"],
                 "exporterVersion": get_version_string()
             },
             required=False
@@ -159,7 +153,7 @@ class glTF2ExportUserExtension:
         registered_hubs_components = get_components_registry()
 
         if component_list.items:
-            extension_name = hubs_config["gltfExtensionName"]
+            extension_name = HUBS_CONFIG["gltfExtensionName"]
             component_data = {}
 
             for component_item in component_list.items:
@@ -170,7 +164,8 @@ class glTF2ExportUserExtension:
                         blender_object, component_class.get_id())
                     data = component.gather(export_settings, blender_object)
                     if hasattr(data, "delayed_gather"):
-                        self.delayed_gathers.append((component_data, component_class.get_name(), data))
+                        self.delayed_gathers.append(
+                            (component_data, component_class.get_name(), data))
                     else:
                         component_data[component_class.get_name()] = data
                 else:
@@ -188,80 +183,11 @@ class glTF2ExportUserExtension:
             self.was_used = True
 
 
-class HubsComponentsExtensionProperties(bpy.types.PropertyGroup):
-    enabled: bpy.props.BoolProperty(
-        name="Export Hubs Components",
-        description='Include this extension in the exported glTF file',
-        default=True
-    )
-    version: IntVectorProperty(size=3)
-
-
-class HubsGLTFExportPanel(bpy.types.Panel):
-
-    bl_idname = "HBA_PT_Export_Panel"
-    bl_label = "Hubs Export Panel"
-    bl_space_type = 'FILE_BROWSER'
-    bl_region_type = 'TOOL_PROPS'
-    bl_label = "Hubs Components"
-    bl_parent_id = "GLTF_PT_export_user_extensions"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    @classmethod
-    def poll(cls, context):
-        sfile = context.space_data
-        operator = sfile.active_operator
-        return operator.bl_idname == "EXPORT_SCENE_OT_gltf"
-
-    def draw_header(self, context):
-        props = bpy.context.scene.HubsComponentsExtensionProperties
-        self.layout.prop(props, 'enabled', text="")
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = True
-        layout.use_property_decorate = False  # No animation.
-
-        props = bpy.context.scene.HubsComponentsExtensionProperties
-        layout.active = props.enabled
-
-        box = layout.box()
-        box.label(text="No options yet")
-
-# called by gltf-blender-io after it has loaded
-
-
-def register_export_panel():
-    try:
-        bpy.utils.register_class(HubsGLTFExportPanel)
-    except Exception:
-        pass
-    return unregister_export_panel
-
-
-def unregister_export_panel():
-    # Since panel is registered on demand, it is possible it is not registered
-    try:
-        bpy.utils.unregister_class(HubsGLTFExportPanel)
-    except Exception:
-        pass
-
-
 def register():
     print("Register GLTF Exporter")
-    register_export_panel()
-    if bpy.app.version < (3, 0, 0):
-        gltf2_blender_export.__gather_gltf = patched_gather_gltf
-    bpy.utils.register_class(HubsComponentsExtensionProperties)
-    bpy.types.Scene.HubsComponentsExtensionProperties = PointerProperty(
-        type=HubsComponentsExtensionProperties)
+    gltf2_blender_export.__gather_gltf = patched_gather_gltf
 
 
 def unregister():
     print("Unregister GLTF Exporter")
-    unregister_export_panel()
-    del bpy.types.Scene.HubsComponentsExtensionProperties
-    bpy.utils.unregister_class(HubsComponentsExtensionProperties)
-    if bpy.app.version < (3, 0, 0):
-        gltf2_blender_export.__gather_gltf = orig_gather_gltf
-    unregister_export_panel()
+    gltf2_blender_export.__gather_gltf = orig_gather_gltf
